requirements(1).txt
streamlit==1.27.2
huggingface-hub==0.16.4
langchain==0.0.315
ctransformers==0.2.27
transformers==4.28.1
torch
faiss-cpu==1.7.4 
sentence-transformers==2.2.2
InstructorEmbedding==1.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1 
instructorembedding==1.0.0
